big refactoring: we keep Nuklear for the final result, OpenGL for training, and we forget Evas
