+++ opengl/one-color-GL.c
@@ -1,3 +1,27 @@
+/******************************************************************************
+ * MIT License                                                                *
+ *                                                                            *
+ * Copyright (c) 2022.  Bernard Tatin                                         *
+ *                                                                            *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell  *
+ * copies of the Software, and to permit persons to whom the Software is      *
+ * furnished to do so, subject to the following conditions:                   *
+ *                                                                            *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.                            *
+ *                                                                            *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR *
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,   *
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE*
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER     *
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.                                                                  *
+ ******************************************************************************/
+
 //
 // Created by bernard on 06/04/2022.
 //
@@ -19,9 +43,9 @@
 int nFrames = 0;
 
 GLWindow mainWindow = {
-    .width = DEFAULT_WIDTH,
-    .height = DEFAULT_HEIGHT,
-    .value = 2
+        .width = DEFAULT_WIDTH,
+        .height = DEFAULT_HEIGHT,
+        .value = 2
 };
 
 
@@ -36,14 +60,15 @@ static inline void unit_quad(void) {
             .width = mainWindow.width,
             .height = mainWindow.height
     };
-    OCRectangle rectangle;
-    scale_image(dimensions, &rectangle);
+    OCRectangle  rectangle;
+    scale_image(dimensions, main_image.ratio, &rectangle);
 
-    quad_vertex(0, 0, rectangle.left,  rectangle.top);
-    quad_vertex(0, 1, rectangle.left,  rectangle.height);
+    quad_vertex(0, 0, rectangle.left, rectangle.top);
+    quad_vertex(0, 1, rectangle.left, rectangle.height);
     quad_vertex(1, 1, rectangle.width, rectangle.height);
     quad_vertex(1, 0, rectangle.width, rectangle.top);
 }
+
 /* Handler for window-repaint event. Called back when the window first appears and
    whenever the window needs to be re-painted. */
 static void displayFunc() {
@@ -79,7 +104,7 @@ static void reshapeFunc(GLsizei new_width, GLsizei new_height) {
 
     // printf("reshape(%d, %d) ", new_width, new_height );
 
-    mainWindow.width = new_width;
+    mainWindow.width  = new_width;
     mainWindow.height = new_height;
     // Set the viewport to cover the new window
     glViewport(0, 0, mainWindow.width, mainWindow.height);
@@ -131,7 +156,7 @@ int main(int argc, char **argv) {
     glutReshapeFunc(reshapeFunc);       // Register callback handler for window re-size event
 
     /* OpenGL 2D generic init */
-//    initGL(mainWindow.width, mainWindow.height);
+    //    initGL(mainWindow.width, mainWindow.height);
 
     /* Initialization of DevIL */
     if (!init_DevIL()) {
